refactor(sync): move sync logic to a dedicated module for better organization
